PerplexicaPython/requirements.txt
# Core dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM
groq>=0.4.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1

# HTTP client
httpx>=0.26.0
requests>=2.31.0

# Search and parsing
beautifulsoup4>=4.12.0
lxml>=5.1.0
duckduckgo-search>=5.0.0
markdown>=3.5.0

# Utilities
tiktoken>=0.5.2
python-dateutil>=2.8.2

# Widgets
# Weather API client (optional, can use requests)
# Stock data
yfinance>=0.2.36

# Math calculations
sympy>=1.12
numexpr>=2.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Code quality
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0

# Logging
loguru>=0.7.2
